Move gui_fit.py into xija as a module
